+++ R/genAssign.R
@@ -8,7 +8,7 @@
 #' 
 #' 0 is the base population.
 #' 
-#' Migrants, or any individuals where both parens are unknown, are assigned to
+#' Migrants, or any individuals where both parents are unknown, are assigned to
 #' generation zero.  If parents of an individual are from two different
 #' generations (e.g., dam = 0 and sire = 1), the individual is assigned to the
 #' generation following the greater of the two parents (e.g., 2 in this
@@ -34,7 +34,9 @@ genAssign <- function(pedigree, ...){
 genAssign.default <- function(pedigree, ...)
 { 
    n <- nrow(pedigree)
-   if(any(apply(pedigree[, 1:3], MARGIN = 2, FUN = function(x){min(x, na.rm = TRUE) < 0}))){
+   numbCols <- which(apply(pedigree[, 1:3], MARGIN = 2, FUN = is.integer) |
+	apply(pedigree[, 1:3], MARGIN = 2, FUN = is.numeric))
+   if(length(numbCols) > 0 && any(apply(pedigree[, numbCols], MARGIN = 2, FUN = function(x){min(x, na.rm = TRUE) < 0}))){
       warning("Negative values in pedigree interpreted as missing values")
       pedigree[pedigree < 0] <- -998
    }