+++ R/makeDomEpi.R
@@ -12,7 +12,7 @@
 #' \code{\link{makeD}}), this function allows for the inverses that are derived
 #' from the D matrix (i.e., D-inverse, AD-inverse, and DD-inverse)to be
 #' constructed at the same time.  This way, the D matrix will only have to be
-#' constructed once for use in the three seperate genetic relatedness inverse
+#' constructed once for use in the three separate genetic relatedness inverse
 #' matrices that depend upon it.  However, using the \code{output} and
 #' \code{invertD} options in different combinations will ensure that only the
 #' desired matrix inverses are constructed.
@@ -64,7 +64,8 @@
 #'   str(DADoutput)
 #' 
 #' @export
-makeDomEpi <- function(pedigree, output = c("AD", "DD", "both"), parallel = FALSE, invertD = FALSE, det = FALSE)
+makeDomEpi <- function(pedigree, output = c("AD", "DD", "both"),
+	parallel = FALSE, invertD = FALSE, det = TRUE)
 {
   type <- match.arg(output)
   Dout <- makeD(pedigree, parallel = parallel, invertD = invertD, returnA = TRUE)